refactor(capabilities): split http client/server into submodule dirs
Turn the flat http/client.rs and http/server.rs files into submodule
directories (http/client/{mod.rs,config.rs} and
http/server/{mod.rs,config.rs}), moving each cap's config struct into
its own config.rs per the parent #2221 config-homing requirement.

- http/client.rs → http/client/mod.rs (via git mv, preserving history).
  HttpConfig + impl Default extracted to http/client/config.rs.
  client/mod.rs adds mod config; + pub use config::HttpConfig and the
  native-gated pub use config::{HttpConfigLayer, HttpOverlay}.
- http/server.rs → http/server/mod.rs (via git mv). HttpServerConfig +
  impl Default extracted to http/server/config.rs. server/mod.rs adds
  mod config; + pub use config::HttpServerConfig and the native-gated
  pub use config::{HttpServerConfigLayer, HttpServerOverlay}.
- http/mod.rs is unchanged — its pub mod client; / pub mod server; now
  resolve to the directories, and all six pub use client::... /
  server::... lines continue to resolve through the new root re-exports.

No behavior change. Full workspace validation (fmt + clippy + nextest
2036/2036 + rustdoc) green.

Closes #2226.
